Rename tests to avoid path too long issues
- Put in-memory tests in their own namespace
