+++ test/Kestrel.InMemory.FunctionalTests/CertificateLoaderTests.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Testing;
 using Microsoft.Extensions.Logging.Testing;
 using Xunit;
 
-namespace Microsoft.AspNetCore.Server.Kestrel.FunctionalTests
+namespace Microsoft.AspNetCore.Server.Kestrel.InMemory.FunctionalTests
 {
     public class CertificateLoaderTests : LoggedTest
     {
