+++ test/Kestrel.InMemory.FunctionalTests/TestTransport/InMemoryConnection.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Adapter.Internal;
 using Microsoft.AspNetCore.Testing;
 
-namespace InMemory.FunctionalTests.TestTransport
+namespace Microsoft.AspNetCore.Server.Kestrel.InMemory.FunctionalTests.TestTransport
 {
     public class InMemoryConnection : StreamBackedTestConnection
     {
