+++ test/Kestrel.Tests/HttpsConnectionAdapterOptionsTest.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.AspNetCore.Server.Kestrel.Https;
 using Xunit;
 
-namespace Microsoft.AspNetCore.Server.Kestrel.FunctionalTests
+namespace Microsoft.AspNetCore.Server.Kestrel.Tests
 {
     public class HttpsConnectionAdapterOptionsTests
     {
