+++ test/Kestrel.Transport.Sockets.BindTests/Kestrel.Transport.Sockets.BindTests.csproj
@@ -1,16 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <AssemblyName>Libuv.AddressRegistrationTests</AssemblyName>
-    <RootNamespace>Libuv.AddressRegistrationTests</RootNamespace>
+    <AssemblyName>Sockets.BindTests</AssemblyName>
+    <RootNamespace>Sockets.BindTests</RootNamespace>
     <TargetFrameworks>$(StandardTestTfms)</TargetFrameworks>
     <ServerGarbageCollection>true</ServerGarbageCollection>
-    <TestGroupName>Libuv.AddressRegistrationTests</TestGroupName>
+    <TestGroupName>Sockets.BindTests</TestGroupName>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\Kestrel.Transport.AddressRegistrationTests\**\*.cs" />
-    <Compile Include="..\Kestrel.Transport.Libuv.FunctionalTests\TransportSelector.cs" />
+    <Compile Include="..\Kestrel.Transport.BindTests\**\*.cs" />
+    <Compile Include="..\Kestrel.Transport.Sockets.FunctionalTests\TransportSelector.cs" />
     <Compile Include="..\shared\*.cs" LinkBase="shared" />
     <Compile Include="..\shared\FunctionalTestHelpers\*.cs" LinkBase="shared\FunctionalTestHelpers" />
     <Compile Include="..\shared\TransportTestHelpers\*.cs" LinkBase="shared\TransportTestHelpers" />
@@ -19,7 +19,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Kestrel\Kestrel.csproj" />
-    <ProjectReference Include="..\..\src\Kestrel.Transport.Libuv\Kestrel.Transport.Libuv.csproj" />
+    <ProjectReference Include="..\..\src\Kestrel.Transport.Sockets\Kestrel.Transport.Sockets.csproj" />
     <ProjectReference Include="..\..\src\Kestrel.Https\Kestrel.Https.csproj" />
   </ItemGroup>
